requirements_all.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.2
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
iniconfig==2.0.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyspark==3.5.4
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
rsa==4.9
SQLAlchemy==2.0.38
tzdata==2025.1
urllib3==2.3.0
